Lista hoteles

+++ lib/src/models/review_model.dart
@@ -4,7 +4,7 @@ class ReviewModel {
   String? idReview;
   String? idHotel;
   String? idUser;
-  int? score;
+  double? score;
   String? comment;
 
   ReviewModel(
